add a network mode and tests for fargate
AWS requires networkMode to be one of [bridge,host,none,awsvpc] or
not present, in which case it shows as <default> in the interface and
uses bridge mode on Linux and host mode on Windows. AWS won't accept
a networkMode field with no value, so if we don't explicitly give a
value in the inputfile, we leave off the networkMode option entirely
in the output.  We're doing no error checking for the values in the
networkMode file, though.